update and optimizing code for calendar page
