+++ src/components/main/calendar/Calendar.tsx
@@ -1,55 +1,23 @@
 "use client"
 
-import { useState, useEffect, useCallback, useMemo } from "react"
-import {
-  ChevronLeft,
-  ChevronRight,
-  Plus,
-  List,
-  Grid,
-  Sun,
-  Moon,
-  Search,
-  ChevronDown,
-  AlertTriangle,
-  AlertCircle,
-  CheckCircle2,
-} from "lucide-react"
-import { Button } from "@/components/ui/button"
-import { cn } from "@/lib/utils"
-import TaskModal from "./task-modal"
+import { useState, useEffect } from "react"
 import { Tabs, TabsContent } from "@/components/ui/calendar/tabs"
-import { Input } from "@/components/ui/input"
-import { ToggleGroup, ToggleGroupItem } from "@/components/ui/calendar/toggle-group"
-import { useTheme } from "next-themes"
-import { motion, AnimatePresence, MotionProps } from "framer-motion"
-import { DropdownMenu, DropdownMenuContent, DropdownMenuItem, DropdownMenuTrigger } from "@/components/ui/dropdown-menu"
 import { useMobile } from "@/hooks/use-mobile"
 import TaskDetailModal from "./task-detail-modal"
-import TaskItem from "./task-item"
+import TaskModal from "./task-modal"
 import { fetchTasks, fetchCategories } from "@/lib/tasks-data"
-import type { Task, Category } from "@/types"
-import { isToday, getDayStatus } from "@/lib/calendar-utils"
-import { HTMLAttributes } from 'react'
-
-type MotionDivProps = MotionProps & HTMLAttributes<HTMLDivElement>
+import { useCalendar } from "@/hooks/use-calendar"
+import CalendarHeader from "./calendar-header"
+import CalendarToolbar from "./calendar-toolbar"
+import MonthView from "./calendar-views/month-view"
+import ListView from "./calendar-views/list-view"
 
 export default function Calendar() {
-  const [currentDate, setCurrentDate] = useState(new Date())
-  const [tasks, setTasks] = useState<Task[]>([])
-  const [categories, setCategories] = useState<Category[]>([])
-  const [isModalOpen, setIsModalOpen] = useState(false)
-  const [selectedDate, setSelectedDate] = useState<string | null>(null)
-  const [view, setView] = useState<"month" | "list">("month")
-  const [searchQuery, setSearchQuery] = useState("")
-  const [selectedCategories, setSelectedCategories] = useState<string[]>([])
-  const [direction, setDirection] = useState<"left" | "right" | null>(null)
-  const [animationKey, setAnimationKey] = useState(0)
-  const [selectedTask, setSelectedTask] = useState<Task | null>(null)
-  const [showCompleted, setShowCompleted] = useState(true)
   const [isLoading, setIsLoading] = useState(true)
-  const { theme, setTheme } = useTheme()
   const isMobile = useMobile()
+
+  // Get all calendar functionality from the custom hook
+  const calendar = useCalendar()
 
   // Fetch tasks and categories on component mount
   useEffect(() => {
@@ -57,8 +25,8 @@ export default function Calendar() {
       setIsLoading(true)
       try {
         const [tasksData, categoriesData] = await Promise.all([fetchTasks(), fetchCategories()])
-        setTasks(tasksData)
-        setCategories(categoriesData)
+        calendar.setTasks(tasksData)
+        calendar.setCategories(categoriesData)
       } catch (error) {
         console.error("Error loading data:", error)
       } finally {
@@ -67,296 +35,13 @@ export default function Calendar() {
     }
 
     loadData()
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [])
-
-  // Get current month and year
-  const currentMonth = currentDate.getMonth()
-  const currentYear = currentDate.getFullYear()
-
-  // Get the first day of the month
-  const firstDayOfMonth = useMemo(() => new Date(currentYear, currentMonth, 1), [currentMonth, currentYear])
-  const startingDayOfWeek = useMemo(() => firstDayOfMonth.getDay(), [firstDayOfMonth])
-
-  // Get the number of days in the month
-  const daysInMonth = useMemo(() => new Date(currentYear, currentMonth + 1, 0).getDate(), [currentMonth, currentYear])
-
-  // Get the name of the month
-  const monthName = useMemo(
-    () => new Intl.DateTimeFormat("en-US", { month: "long" }).format(currentDate),
-    [currentDate],
-  )
-
-
-  // Toggle task completion status
-  const toggleTaskCompletion = useCallback((taskId: number) => {
-    setTasks((prevTasks) =>
-      prevTasks.map((task) => (task.id === taskId ? { ...task, completed: !task.completed } : task)),
-    )
-  }, [])
-
-  // Filter tasks based on search query, selected categories, and completion status
-  const filterTasks = useCallback(
-    (taskList: Task[]) => {
-      let filtered = [...taskList]
-
-      // Filter by completion status if needed
-      if (!showCompleted) {
-        filtered = filtered.filter((task) => !task.completed)
-      }
-
-      if (searchQuery) {
-        const query = searchQuery.toLowerCase()
-        filtered = filtered.filter(
-          (task) =>
-            task.title.toLowerCase().includes(query) ||
-            (task.description && task.description.toLowerCase().includes(query)),
-        )
-      }
-
-      if (selectedCategories.length > 0) {
-        filtered = filtered.filter((task) => selectedCategories.includes(task.category))
-      }
-
-      return filtered
-    },
-    [searchQuery, selectedCategories, showCompleted],
-  )
-
-  // Navigate to previous month with animation
-  const prevMonth = useCallback(() => {
-    setDirection("right")
-    setTimeout(() => {
-      setCurrentDate((prev) => {
-        const newDate = new Date(prev)
-        newDate.setMonth(prev.getMonth() - 1)
-        return newDate
-      })
-      setAnimationKey((prev) => prev + 1)
-    }, 50)
-  }, [])
-
-  // Navigate to next month with animation
-  const nextMonth = useCallback(() => {
-    setDirection("left")
-    setTimeout(() => {
-      setCurrentDate((prev) => {
-        const newDate = new Date(prev)
-        newDate.setMonth(prev.getMonth() + 1)
-        return newDate
-      })
-      setAnimationKey((prev) => prev + 1)
-    }, 50)
-  }, [])
-
-  // Go to today
-  const goToToday = useCallback(() => {
-    const today = new Date()
-    if (today.getMonth() !== currentMonth || today.getFullYear() !== currentYear) {
-      if (today.getTime() < currentDate.getTime()) {
-        setDirection("right")
-      } else {
-        setDirection("left")
-      }
-      setTimeout(() => {
-        setCurrentDate(today)
-        setAnimationKey((prev) => prev + 1)
-      }, 50)
-    }
-  }, [currentDate, currentMonth, currentYear])
-
-  // Change year
-  const changeYear = useCallback(
-    (year: number) => {
-      if (year !== currentYear) {
-        setDirection(year < currentYear ? "right" : "left")
-        setTimeout(() => {
-          setCurrentDate((prev) => {
-            const newDate = new Date(prev)
-            newDate.setFullYear(year)
-            return newDate
-          })
-          setAnimationKey((prev) => prev + 1)
-        }, 50)
-      }
-    },
-    [currentYear],
-  )
-
-  // Change month
-  const changeMonth = useCallback(
-    (monthIndex: number) => {
-      if (monthIndex !== currentMonth) {
-        setDirection(monthIndex < currentMonth ? "right" : "left")
-        setTimeout(() => {
-          setCurrentDate((prev) => {
-            const newDate = new Date(prev)
-            newDate.setMonth(monthIndex)
-            return newDate
-          })
-          setAnimationKey((prev) => prev + 1)
-        }, 50)
-      }
-    },
-    [currentMonth],
-  )
-
-  // Generate array of years for the year picker (current year ±10 years)
-  const getYearOptions = useMemo(() => {
-    const currentYear = new Date().getFullYear()
-    return Array.from({ length: 21 }, (_, i) => currentYear - 10 + i)
-  }, [])
-
-  // Reset animation direction after animation completes
-  useEffect(() => {
-    const timer = setTimeout(() => {
-      setDirection(null)
-    }, 500)
-    return () => clearTimeout(timer)
-  }, [animationKey])
-
-  // Create calendar days array
-  const calendarDays = useMemo(() => {
-    const days = []
-
-    // Add empty cells for days before the first day of the month
-    for (let i = 0; i < startingDayOfWeek; i++) {
-      days.push(null)
-    }
-
-    // Add days of the month
-    for (let day = 1; day <= daysInMonth; day++) {
-      days.push(day)
-    }
-
-    return days
-  }, [daysInMonth, startingDayOfWeek])
-
-  // Get tasks for a specific day
-  const getTasksForDay = useCallback(
-    (day: number) => {
-      const date = `${currentYear}-${String(currentMonth + 1).padStart(2, "0")}-${String(day).padStart(2, "0")}`
-      return filterTasks(
-        tasks.filter((task) => {
-          // Extract date part from dueDate (YYYY-MM-DD)
-          const taskDate = task.dueDate.split("T")[0]
-          return taskDate === date
-        }),
-      )
-    },
-    [currentMonth, currentYear, tasks, filterTasks],
-  )
-
-  // Open modal to add a task for a specific day
-  const openAddTaskModal = useCallback(
-    (day: number) => {
-      const date = `${currentYear}-${String(currentMonth + 1).padStart(2, "0")}-${String(day).padStart(2, "0")}`
-      setSelectedDate(date)
-      setIsModalOpen(true)
-    },
-    [currentMonth, currentYear],
-  )
-
-  // Add a new task
-  const addTask = useCallback(
-    (task: Omit<Task, "id" | "createdAt" | "starred">) => {
-      const newTask: Task = {
-        ...task,
-        id: tasks.length > 0 ? Math.max(...tasks.map((t) => t.id)) + 1 : 1,
-        createdAt: new Date().toISOString(),
-        starred: false,
-      }
-
-      setTasks((prevTasks) => [...prevTasks, newTask])
-      setIsModalOpen(false)
-    },
-    [tasks],
-  )
-
-  // Get all tasks for the current month
-  const getMonthTasks = useCallback(() => {
-    const monthStart = new Date(currentYear, currentMonth, 1)
-    const monthEnd = new Date(currentYear, currentMonth + 1, 0)
-
-    return filterTasks(
-      tasks.filter((task) => {
-        // Extract date part from dueDate (YYYY-MM-DD)
-        const taskDate = new Date(task.dueDate.split("T")[0])
-        return taskDate >= monthStart && taskDate <= monthEnd
-      }),
-    ).sort((a, b) => new Date(a.dueDate).getTime() - new Date(b.dueDate).getTime())
-  }, [currentMonth, currentYear, filterTasks, tasks])
-
-  // Toggle category selection
-  const toggleCategory = useCallback((category: string) => {
-    setSelectedCategories((prev) =>
-      prev.includes(category) ? prev.filter((c) => c !== category) : [...prev, category],
-    )
-  }, [])
-
-  // Clear all selected categories
-  const clearCategoryFilters = useCallback(() => {
-    setSelectedCategories([])
-  }, [])
-
-  // Get all unique categories from tasks
-  const allCategories = useMemo(() => Array.from(new Set(tasks.map((task) => task.category))), [tasks])
 
   // Get day names
   const dayNames = ["Sun", "Mon", "Tue", "Wed", "Thu", "Fri", "Sat"]
   // For mobile, use shorter day names
   const shortDayNames = ["S", "M", "T", "W", "T", "F", "S"]
-
-  // Open task detail modal
-  const openTaskDetail = useCallback((task: Task) => {
-    setSelectedTask(task)
-  }, [])
-
-  // Close task detail modal
-  const closeTaskDetail = useCallback(() => {
-    setSelectedTask(null)
-  }, [])
-
-  // Toggle theme
-  const toggleTheme = useCallback(() => {
-    setTheme(theme === "dark" ? "light" : "dark")
-  }, [theme, setTheme])
-
-  // Toggle showing completed tasks
-  const toggleShowCompleted = useCallback(() => {
-    setShowCompleted((prev) => !prev)
-  }, [])
-
-  // Animation variants for month transitions
-  const variants = {
-    enter: (direction: "left" | "right" | null) => ({
-      x: direction === "left" ? 1000 : direction === "right" ? -1000 : 0,
-      opacity: 0,
-    }),
-    center: {
-      x: 0,
-      opacity: 1,
-    },
-    exit: (direction: "left" | "right" | null) => ({
-      x: direction === "left" ? -1000 : direction === "right" ? 1000 : 0,
-      opacity: 0,
-    }),
-  }
-
-  // Animation for month name
-  const monthVariants = {
-    initial: { y: -20, opacity: 0 },
-    animate: { y: 0, opacity: 1 },
-    exit: { y: 20, opacity: 0 },
-  }
-
-  // Ensure theme is properly applied on initial load
-  useEffect(() => {
-    const savedTheme = localStorage.getItem("calendar-theme")
-    if (savedTheme) {
-      setTheme(savedTheme)
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [])
 
   if (isLoading) {
     return (
@@ -369,385 +54,88 @@ export default function Calendar() {
   return (
     <div className="bg-white dark:bg-gray-950 rounded-2xl shadow-xl overflow-hidden border border-gray-200 dark:border-gray-800 transition-colors duration-300 flex flex-col h-full overflow-y-auto">
       {/* Calendar header */}
-      <div className="p-4 flex flex-col sm:flex-row items-center justify-between bg-gradient-to-r from-indigo-500 to-purple-600 text-white rounded-t-2xl">
-        <div className="flex items-center mb-2 sm:mb-0">
-          <div className="flex items-center">
-            {/* Month Picker */}
-            <DropdownMenu>
-              <DropdownMenuTrigger asChild>
-                <Button variant="ghost" className="font-bold text-xl hover:bg-white/20 rounded-xl">
-                  <AnimatePresence mode="wait">
-                    <motion.span
-                      key={monthName}
-                      initial="initial"
-                      animate="animate"
-                      exit="exit"
-                      variants={monthVariants}
-                      transition={{ duration: 0.3 }}
-                    >
-                      {isMobile ? monthName.substring(0, 3) : monthName}
-                    </motion.span>
-                  </AnimatePresence>
-                  <ChevronDown className="ml-1 h-4 w-4" />
-                </Button>
-              </DropdownMenuTrigger>
-              <DropdownMenuContent align="start" className="rounded-xl">
-                {Array.from({ length: 12 }, (_, i) => {
-                  const date = new Date(currentYear, i, 1)
-                  const monthName = new Intl.DateTimeFormat("en-US", { month: "long" }).format(date)
-                  return (
-                    <DropdownMenuItem
-                      key={i}
-                      onClick={() => changeMonth(i)}
-                      className={cn(
-                        "cursor-pointer rounded-lg",
-                        i === currentMonth &&
-                          "bg-indigo-100 text-indigo-900 dark:bg-indigo-900 dark:text-indigo-100 font-medium",
-                      )}
-                    >
-                      {monthName}
-                    </DropdownMenuItem>
-                  )
-                })}
-              </DropdownMenuContent>
-            </DropdownMenu>
-
-            {/* Year Picker */}
-            <DropdownMenu>
-              <DropdownMenuTrigger asChild>
-                <Button variant="ghost" className="ml-1 font-bold text-xl hover:bg-white/20 rounded-xl">
-                  <AnimatePresence mode="wait">
-                    <motion.span
-                      key={currentYear}
-                      initial="initial"
-                      animate="animate"
-                      exit="exit"
-                      variants={monthVariants}
-                      transition={{ duration: 0.3 }}
-                    >
-                      {currentYear}
-                    </motion.span>
-                  </AnimatePresence>
-                  <ChevronDown className="ml-1 h-4 w-4" />
-                </Button>
-              </DropdownMenuTrigger>
-              <DropdownMenuContent align="start" className="h-[300px] overflow-y-auto rounded-xl">
-                {getYearOptions.map((year) => (
-                  <DropdownMenuItem
-                    key={year}
-                    onClick={() => changeYear(year)}
-                    className={cn(
-                      "cursor-pointer rounded-lg",
-                      year === currentYear &&
-                        "bg-indigo-100 text-indigo-900 dark:bg-indigo-900 dark:text-indigo-100 font-medium",
-                    )}
-                  >
-                    {year}
-                  </DropdownMenuItem>
-                ))}
-              </DropdownMenuContent>
-            </DropdownMenu>
-          </div>
-        </div>
-
-        <div className="flex items-center space-x-2">
-          <Button
-            variant="outline"
-            size="sm"
-            onClick={goToToday}
-            className="bg-white/20 hover:bg-white/30 text-white border-white/40 rounded-xl"
-          >
-            Today
-          </Button>
-          <Button variant="ghost" size="icon" onClick={prevMonth} className="hover:bg-white/20 text-white rounded-xl">
-            <ChevronLeft className="h-5 w-5" />
-          </Button>
-          <Button variant="ghost" size="icon" onClick={nextMonth} className="hover:bg-white/20 text-white rounded-xl">
-            <ChevronRight className="h-5 w-5" />
-          </Button>
-          <Button
-            variant="ghost"
-            size="icon"
-            onClick={toggleTheme}
-            className="hover:bg-white/20 text-white rounded-xl relative group"
-            aria-label={theme === "dark" ? "Switch to light mode" : "Switch to dark mode"}
-          >
-            {theme === "dark" ? <Sun className="h-5 w-5" /> : <Moon className="h-5 w-5" />}
-            <span className="sr-only">{theme === "dark" ? "Light mode" : "Dark mode"}</span>
-            <div className="absolute -bottom-10 left-1/2 transform -translate-x-1/2 px-2 py-1 bg-black/80 text-white text-xs rounded-md opacity-0 group-hover:opacity-100 transition-opacity pointer-events-none whitespace-nowrap">
-              {theme === "dark" ? "Switch to light mode" : "Switch to dark mode"}
-            </div>
-          </Button>
-        </div>
-      </div>
+      <CalendarHeader
+        monthName={calendar.monthName}
+        currentYear={calendar.currentYear}
+        currentMonth={calendar.currentMonth}
+        isMobile={isMobile}
+        goToToday={calendar.goToToday}
+        prevMonth={calendar.prevMonth}
+        nextMonth={calendar.nextMonth}
+        changeMonth={calendar.changeMonth}
+        changeYear={calendar.changeYear}
+        getYearOptions={calendar.getYearOptions}
+      />
 
       {/* Toolbar */}
-      <div className="p-2 bg-gray-50 dark:bg-gray-900 border-b border-gray-200 dark:border-gray-800 flex flex-col sm:flex-row items-center justify-between gap-2">
-        <div className="flex items-center w-full sm:w-auto">
-          <div className="relative w-full sm:w-64">
-            <Search className="absolute left-2 top-2.5 h-4 w-4 text-muted-foreground" />
-            <Input
-              placeholder="Search tasks..."
-              className="pl-8 rounded-xl"
-              value={searchQuery}
-              onChange={(e) => setSearchQuery(e.target.value)}
-            />
-          </div>
-        </div>
-
-        <div className="flex items-center gap-2 w-full sm:w-auto justify-between sm:justify-end">
-          <DropdownMenu>
-            <DropdownMenuTrigger asChild>
-              <Button variant="outline" size="sm" className="rounded-lg flex items-center gap-1">
-                Categories
-                <ChevronDown className="h-3 w-3 ml-1" />
-              </Button>
-            </DropdownMenuTrigger>
-            <DropdownMenuContent align="start" className="max-h-[300px] overflow-y-auto rounded-xl">
-              {categories.map((category) => (
-                <DropdownMenuItem
-                  key={category.name}
-                  className={cn(
-                    "cursor-pointer rounded-lg flex items-center gap-2",
-                    selectedCategories.includes(category.name) &&
-                      "bg-indigo-100 text-indigo-900 dark:bg-indigo-900 dark:text-indigo-100 font-medium",
-                  )}
-                  onClick={() => toggleCategory(category.name)}
-                >
-                  <div className="w-3 h-3 rounded-full" style={{ backgroundColor: category.color }} />
-                  {category.name}
-                  {selectedCategories.includes(category.name) && <CheckCircle2 className="h-4 w-4 ml-auto" />}
-                </DropdownMenuItem>
-              ))}
-            </DropdownMenuContent>
-          </DropdownMenu>
-          {selectedCategories.length > 0 && (
-            <Button variant="ghost" size="sm" onClick={clearCategoryFilters} className="text-xs rounded-lg ml-1">
-              Clear filters ({selectedCategories.length})
-            </Button>
-          )}
-
-          <Button
-            variant="outline"
-            size="sm"
-            onClick={toggleShowCompleted}
-            className={cn(
-              "text-xs rounded-lg flex items-center gap-1",
-              showCompleted ? "bg-green-100 text-green-800 dark:bg-green-900/30 dark:text-green-300" : "",
-            )}
-          >
-            {showCompleted ? <CheckCircle2 className="h-3 w-3" /> : null}
-            {showCompleted ? "Hide" : "Show"} completed
-          </Button>
-
-          <ToggleGroup
-            type="single"
-            value={view}
-            onValueChange={(value) => value && setView(value as "month" | "list")}
-          >
-            <ToggleGroupItem value="month" aria-label="Month view" className="rounded-l-xl">
-              <Grid className="h-4 w-4" />
-            </ToggleGroupItem>
-            <ToggleGroupItem value="list" aria-label="List view" className="rounded-r-xl">
-              <List className="h-4 w-4" />
-            </ToggleGroupItem>
-          </ToggleGroup>
-        </div>
-      </div>
+      <CalendarToolbar
+        searchQuery={calendar.searchQuery}
+        setSearchQuery={calendar.setSearchQuery}
+        view={calendar.view}
+        setView={calendar.setView}
+        categories={calendar.categories}
+        selectedCategories={calendar.selectedCategories}
+        toggleCategory={calendar.toggleCategory}
+        clearCategoryFilters={calendar.clearCategoryFilters}
+        showCompleted={calendar.showCompleted}
+        toggleShowCompleted={calendar.toggleShowCompleted}
+      />
 
       {/* Tabs for different views */}
-      <Tabs value={view} className="w-full" onValueChange={(value) => setView(value as "month" | "list")}>
+      <Tabs
+        value={calendar.view}
+        className="w-full"
+        onValueChange={(value) => calendar.setView(value as "month" | "list")}
+      >
         {/* Month View */}
         <TabsContent value="month" className="m-0 overflow-hidden">
-          {/* Day names */}
-          <div className="grid grid-cols-7 bg-gray-100 dark:bg-gray-900">
-            {(isMobile ? shortDayNames : dayNames).map((day, index) => (
-              <div key={index} className="py-2 text-center text-sm font-medium text-gray-500 dark:text-gray-400">
-                {day}
-              </div>
-            ))}
-          </div>
-
-          {/* Calendar grid with animation */}
-          <AnimatePresence initial={false} custom={direction} mode="wait">
-            <motion.div
-              key={animationKey}
-              custom={direction}
-              variants={variants}
-              initial="enter"
-              animate="center"
-              exit="exit"
-              transition={{
-                x: { type: "spring", stiffness: 300, damping: 30 },
-                opacity: { duration: 0.2 },
-              }}
-              className="grid grid-cols-7 gap-1 p-1 bg-gray-200 dark:bg-gray-800"
-              {...({} as MotionDivProps)}
-            >
-              {calendarDays.map((day, index) => {
-                if (day === null) {
-                  return (
-                    <div
-                      key={`empty-${index}`}
-                      className="bg-gray-50 dark:bg-gray-900 h-24 sm:h-28 md:h-32 rounded-xl"
-                    />
-                  )
-                }
-
-                const dayTasks = getTasksForDay(day)
-                const today = isToday(day, currentMonth, currentYear)
-                const dayStatus = getDayStatus(dayTasks)
-
-                return (
-                  <motion.div
-                    key={`day-${day}`}
-                    initial={{ opacity: 0, scale: 0.9 }}
-                    animate={{ opacity: 1, scale: 1 }}
-                    transition={{ duration: 0.2, delay: index * 0.01 }}
-                    className={cn(
-                      "bg-white dark:bg-gray-950 h-24 sm:h-28 md:h-32 p-2 relative transition-all duration-200 group rounded-xl",
-                      today &&
-                        "ring-2 ring-indigo-500 ring-inset shadow-[0_0_15px_rgba(99,102,241,0.5)] dark:shadow-[0_0_15px_rgba(99,102,241,0.3)] z-10",
-                      dayStatus === "overdue" && "ring-2 ring-red-500 ring-inset",
-                      dayStatus === "approaching" && "ring-2 ring-amber-500 ring-inset",
-                      dayStatus === "completed" && "ring-2 ring-green-500 ring-inset",
-                      "hover:bg-gray-50 dark:hover:bg-gray-900",
-                    )}
-                    {...({} as MotionDivProps)}
-                  >
-                    <div className="flex justify-between items-start">
-                      <div
-                        className={cn(
-                          "relative",
-                          today &&
-                            "after:content-[''] after:absolute after:top-[-4px] after:left-[-4px] after:right-[-4px] after:bottom-[-4px] after:bg-indigo-500/20 after:rounded-full after:animate-pulse",
-                        )}
-                      >
-                        <span
-                          className={cn(
-                            "inline-flex h-7 w-7 items-center justify-center rounded-full text-sm relative z-10",
-                            today ? "bg-indigo-500 text-white font-bold shadow-md" : "text-gray-700 dark:text-gray-300",
-                          )}
-                        >
-                          {day}
-                        </span>
-                      </div>
-                    </div>
-
-                    <div className="relative h-[calc(100%-28px)]">
-                      <div className="absolute inset-0 overflow-y-auto space-y-1 pr-1 pt-1 custom-scrollbar">
-                        {dayTasks.map((task) => {
-                          return (
-                            <TaskItem
-                              key={task.id}
-                              task={task}
-                              openTaskDetail={openTaskDetail}
-                              toggleTaskCompletion={toggleTaskCompletion}
-                              view="month"
-                            />
-                          )
-                        })}
-                      </div>
-
-                      {/* Add task button at bottom right */}
-                      <Button
-                        variant="ghost"
-                        size="icon"
-                        className="h-6 w-6 opacity-0 group-hover:opacity-100 hover:bg-gray-100 dark:hover:bg-gray-800 rounded-lg absolute bottom-0 right-0"
-                        onClick={() => openAddTaskModal(day)}
-                      >
-                        <Plus className="h-4 w-4" />
-                      </Button>
-
-                      {/* Show scroll indicator if there are more tasks than can fit */}
-                      {dayTasks.length > 3 && (
-                        <div className="absolute bottom-0 left-0 right-0 h-4 bg-gradient-to-t from-white dark:from-gray-950 to-transparent pointer-events-none" />
-                      )}
-                    </div>
-
-                    {/* Status indicators - positioned differently to avoid overlap */}
-                    <div className="absolute top-1 right-1 flex items-center gap-1">
-                      {dayStatus === "overdue" && <AlertCircle className="h-4 w-4 text-red-600 dark:text-red-400" />}
-
-                      {dayStatus === "approaching" && (
-                        <motion.div
-                          animate={{ opacity: [0.6, 1, 0.6] }}
-                          transition={{ duration: 1.5, repeat: Number.POSITIVE_INFINITY }}
-                        >
-                          <AlertTriangle className="h-4 w-4 text-amber-600 dark:text-amber-400" />
-                        </motion.div>
-                      )}
-
-                      {dayStatus === "completed" && (
-                        <CheckCircle2 className="h-4 w-4 text-green-600 dark:text-green-400" />
-                      )}
-                    </div>
-                  </motion.div>
-                )
-              })}
-            </motion.div>
-          </AnimatePresence>
+          <MonthView
+            calendarDays={calendar.calendarDays}
+            currentMonth={calendar.currentMonth}
+            currentYear={calendar.currentYear}
+            direction={calendar.direction}
+            animationKey={calendar.animationKey}
+            getTasksForDay={calendar.getTasksForDay}
+            openAddTaskModal={calendar.openAddTaskModal}
+            openTaskDetail={calendar.openTaskDetail}
+            toggleTaskCompletion={calendar.toggleTaskCompletion}
+            categoryColorMap={calendar.categoryColorMap}
+            dayNames={isMobile ? shortDayNames : dayNames}
+            isMobile={isMobile}
+          />
         </TabsContent>
 
         {/* List View */}
         <TabsContent value="list" className="m-0">
-          <AnimatePresence initial={false} custom={direction} mode="wait">
-            <motion.div
-              key={`list-${animationKey}`}
-              custom={direction}
-              variants={variants}
-              initial="enter"
-              animate="center"
-              exit="exit"
-              transition={{
-                x: { type: "spring", stiffness: 300, damping: 30 },
-                opacity: { duration: 0.2 },
-              }}
-              className="p-4 space-y-4"
-              {...({} as MotionDivProps)}
-            >
-              {getMonthTasks().length > 0 ? (
-                getMonthTasks().map((task) => {
-                  const isTaskToday = new Date(task.dueDate.split("T")[0]).toDateString() === new Date().toDateString()
-
-                  return (
-                    <TaskItem
-                      key={task.id}
-                      task={task}
-                      openTaskDetail={openTaskDetail}
-                      toggleTaskCompletion={toggleTaskCompletion}
-                      view="list"
-                      isTaskToday={isTaskToday}
-                    />
-                  )
-                })
-              ) : (
-                <div className="text-center py-8 text-gray-500 dark:text-gray-400">
-                  {showCompleted ? "No tasks found for this month" : "No incomplete tasks found for this month"}
-                </div>
-              )}
-            </motion.div>
-          </AnimatePresence>
+          <ListView
+            getMonthTasks={calendar.getMonthTasks}
+            direction={calendar.direction}
+            animationKey={calendar.animationKey}
+            openTaskDetail={calendar.openTaskDetail}
+            toggleTaskCompletion={calendar.toggleTaskCompletion}
+            categoryColorMap={calendar.categoryColorMap}
+            showCompleted={calendar.showCompleted}
+          />
         </TabsContent>
       </Tabs>
 
       {/* Task modal */}
-      {isModalOpen && (
+      {calendar.isModalOpen && (
         <TaskModal
-          isOpen={isModalOpen}
-          onClose={() => setIsModalOpen(false)}
-          onAddTask={addTask}
-          selectedDate={selectedDate}
-          categories={allCategories}
+          isOpen={calendar.isModalOpen}
+          onClose={() => calendar.setIsModalOpen(false)}
+          onAddTask={calendar.addTask}
+          selectedDate={calendar.selectedDate}
+          categories={calendar.allCategories}
         />
       )}
 
       {/* Task detail modal */}
       <TaskDetailModal
-        task={selectedTask}
-        isOpen={!!selectedTask}
-        onClose={closeTaskDetail}
-        toggleTaskCompletion={toggleTaskCompletion}
+        task={calendar.selectedTask}
+        isOpen={!!calendar.selectedTask}
+        onClose={calendar.closeTaskDetail}
+        toggleTaskCompletion={calendar.toggleTaskCompletion}
       />
     </div>
   )
